pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "headless-curator"
version = "1.0.0"
description = "Personalized Apple Music playlist generator that learns from listening behavior"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Dave", email = "dave@example.com" }
]
keywords = ["apple-music", "playlist", "curator", "music-discovery"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio",
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "structlog>=24.1.0",
    "apscheduler>=3.10.4",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "keyring>=25.0.0",
    "pyjwt[crypto]>=2.8.0",
    "aiosqlite>=0.19.0",
    "sqlalchemy[asyncio]>=2.0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
curator = "src.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
